GeometryProcessing, Mesh Operations.  Fix implicit type conversion warnings.
#preflight 62bb1a76727d9bb33ba7d7c1
#rb Jimmy.Andrews

[CL 21007984 by David Hill in ue5-main branch]

+++ Engine/Plugins/Runtime/GeometryProcessing/Source/DynamicMesh/Private/Operations/ExtrudeMesh.cpp
@@ -171,7 +171,7 @@ bool FExtrudeMesh::ApplyExtrude(FExtrusionInfo& Region, FMeshNormals* UseNormals
 
 					if (k > 0)
 					{
-						AccumUVTranslation += Distance(Mesh->GetVertex(BaseLoop.Vertices[k]), Mesh->GetVertex(BaseLoop.Vertices[k - 1]));
+						AccumUVTranslation +=(float) Distance(Mesh->GetVertex(BaseLoop.Vertices[k]), Mesh->GetVertex(BaseLoop.Vertices[k - 1]));
 					}
 
 					// translate horizontally such that vertical spans are adjacent in UV space (so textures tile/wrap properly)
